+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/member/domain/model/Gender.java
@@ -1,0 +1,6 @@
+package com.nextdoor.nextdoor.domain.member.domain.model;
+
+public enum Gender {
+
+    MALE, FEMALE
+}
