--- backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/member/enums/Gender.java
@@ -1,6 +1,0 @@
-package com.nextdoor.nextdoor.domain.member.enums;
-
-public enum Gender {
-
-    MALE, FEMALE
-}
