+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/member/presentation/dto/request/MemberExtraInfoSaveRequestDto.java
@@ -1,6 +1,6 @@
 package com.nextdoor.nextdoor.domain.member.presentation.dto.request;
 
-import com.nextdoor.nextdoor.domain.member.enums.Gender;
+import com.nextdoor.nextdoor.domain.member.domain.model.Gender;
 import lombok.Getter;
 
 @Getter
